+++ CHANGELOG.md
@@ -16,8 +16,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 - `trusted-replace-argument` scriptlet [#405].
 
+### Fixed
+
+- issue with `TrustedScriptURL` in `prevent-element-src-loading` scriptlet [#514].
+
 [Unreleased]: https://github.com/AdguardTeam/Scriptlets/compare/v2.2.8...HEAD
 [#405]: https://github.com/AdguardTeam/Scriptlets/issues/405
+[#514]: https://github.com/AdguardTeam/Scriptlets/issues/514
 
 ## [v2.2.8] - 2025-07-08